States cleanup (#480)
* Remove copyright header

* Move outputting HoI4::State to the outputter module

* Move outputting HoI4States to the outputter module

* Clean up DefaultState a little

* Clean up HoI4::State

* Better name state categories

* Remove a 'using namespace std'

* Remove need for HoI4::State to return the source state

* Don't required stored Vic2 state for infrastructure calculations.

* Only truncate average rail level at the last moment.

* Don't require stored Vic2 state to convert VPs.

* No longer require stored Vic2 state to convert naval bases

* Remove need for stored Vic2 state when converting controlled provinces

* Remove need for stored Vic2 state when converting manpower

* Store actually needed data instead of the Vic2 state.

* No longer store Vic2 world in HoI4States

* HoI4States -> HoI4::States

* Store States as a unique_ptr, not a raw pointer.

* PVS Studio recommendations

* Add states outputter to Linux build

+++ Vic2ToHoI4/Source/V2World/State.cpp
@@ -186,9 +186,9 @@ int Vic2::State::getPopulation() const
 }
 
 
-int Vic2::State::getAverageRailLevel() const
+float Vic2::State::getAverageRailLevel() const
 {
-	int totalRailLevel = 0;
+	float totalRailLevel = 0;
 	for (auto province: provinces)
 	{
 		totalRailLevel += province->getRailLevel();
@@ -200,6 +200,6 @@ int Vic2::State::getAverageRailLevel() const
 	}
 	else
 	{
-		return 0;
+		return 0.0f;
 	}
 }
